Clean and summarize payment data
Clean payment data, and summarize it to the worker level. Meaning changing the level from payment-worker to only worker by adding all payments and taking as the completion date the min of the payment dates, also considering that if a payment was deferred, the worker overall had a deferred payment.