Add comparecencias and reduce to first hearing data
Hearings features were obviously at the request-hearing-worker level. As we previously filtered to ensure there was only one worker, the level is now that of the request-hearing. However, only the first hearing is assigned in a quasi-random manner, so the data set is reduced to information from the first hearing. Additionally, data related to the parties' appearance was added.